+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@
 * #97 "bootique-jetty-junit5" - a test module for Jetty 
 * #98 "any" port - selecting port dynamically
 * #99 SSL context factory class should be changed to comply with recent versions of Jetty
+* #102 Upgrade to Jetty 9.4.31.v20200723
 
 ## 1.1